move io for qlat, channel, waterbody to nhd_io
